bot/requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.3 ; python_version >= '3.6'
aiosignal==1.2.0 ; python_version >= '3.6'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.1.0 ; python_version >= '3.5'
certifi==2022.9.24 ; python_version >= '3.6'
chardet==4.0.0
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
discord.py==2.0.1
frozenlist==1.3.1 ; python_version >= '3.7'
greenlet==1.1.3.post0 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
idna==3.4 ; python_version >= '3.5'
multidict==6.0.2 ; python_version >= '3.7'
requests==2.28.1
sqlalchemy==1.4.42
typing-extensions==4.3.0
urllib3==1.26.12 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
yarl==1.8.1 ; python_version >= '3.7'
